+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [Display of _unknown_ diagnostics counters](https://github.com/christophe-calmejane/Hive/issues/131)
 - Connection matrix displays entities to which Hive is not registered for unsolicited notifications (red label)
 - Displaying _Milan Info_ if available, even if device has been deemed not _Milan Compatible_
+- New statistics counter: Loss of unsolicited notifications for Milan devices
 
 ### Changed
 - Connection matrix _Redundant Intersection_ now indicates _Media Locked_ status if both streams are connected/bound but one has _Interface Down_ status